fetch all recipes on the dashboard
- they are not yet tied to a user. you get all recipes in the db
- add CategoryFilter.tsx component to handle client side events and state
- add loading.tsx
