+++ src/app/dashboard/loading.tsx
@@ -1,0 +1,3 @@
+export default function Loading() {
+  return <div>loading...</div>;
+}
